+++ .env
@@ -29,7 +29,7 @@ WP_ADMIN_PASSWORD=password
 
 WP_DB_PORT=4306
 
-# The databse is served from the Docker `db` container.
+# The database is served from the Docker `db` container.
 # See `dev/docker/ci-compose.yml` for more information.
 WP_TABLE_PREFIX=wp_
 WP_DB_HOST=127.0.0.1:4306
@@ -37,7 +37,7 @@ WP_DB_NAME=wordpress
 WP_DB_USER=root
 WP_DB_PASSWORD=
 
-# The test databse is served from the Docker `db` container.
+# The test database is served from the Docker `db` container.
 # See `dev/docker/ci-compose.yml` for more information.
 WP_TEST_DB_HOST=127.0.0.1:4306
 WP_TEST_DB_NAME=test
